Make Lane an integer. Use the samplesheet to identify fastq files created by bcl2fastq

+++ pipeline-executor/src/main/scala/pipelines/model/AlignmentSummaryMetrics.scala
@@ -54,7 +54,7 @@ object AlignmentSummaryMetrics {
       }
 
       def laneFromReadGroup(readGroup: String): Lane =
-        Lane(readGroup.split("\\.").last)
+        Lane(readGroup.split("\\.").last.toInt)
 
       object H {
         val rg = "READ_GROUP"
